Basic TF-IDF bias detection model
At the moment this does not do anything useful, but it possibly lays down a foundation which we can build on. Populate the "left" "right" and "neutral" folders with articles from known biased sources and then run the script in the .ipynb file to get a list of the "most left" and "most right" words as identified in the text.